requirements.txt
Django==4.1.3
aiogram==2.21
djangorestframework==3.13.1
environs==8.0.0
